pyproject.toml
[tool.poetry]
name = "us-elections"
version = "0.1.0"
description = ""
authors = []
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
torch = "2.3.1"
transformers = "^4.45.2"
pandas = "^2.2.3"
matplotlib = "^3.9.2"
geopandas = "^1.0.1"
ipykernel = "^6.29.5"
python-dotenv = "^1.0.1"
accelerate = "^1.0.1"
lm-eval = "^0.4.5"
autoawq = "^0.2.6"
xlsxwriter = "^3.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
